feat(web): create healthcheck endpoint
Stream the HTML using a custom abstraction (`Page`) on top of Rocket's
`stream` macro and `maud`.

Unfortunately no template library has support for streaming. All that
use macros to compile the markup to Rust code concatenate the fragments
into a a string. From them, `maud` seems to be the one that integrates
most ergonomically with the `select` macro.